server/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
comm==0.1.2
cryptography==37.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.18
django-adaptors==0.2.5
django-allauth==0.50.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
django-extensions==3.1.5
django-imagekit==4.1.0
django-rest-auth==0.9.5
django-seed==0.3.1
django-sslserver==0.22
djangorestframework==3.13.1
entrypoints==0.3
executing==0.8.3
Faker==13.3.4
fastjsonschema==2.16.2
finance-datareader==0.9.31
fonttools==4.31.2
fqdn==1.5.1
idna==3.3
image==1.5.33
importlib-metadata==6.0.0
ipykernel==6.7.0
ipython==8.3.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.0.3
jsonpointer==2.3
jsonschema==4.4.0
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
ngrok==0.7.0
notebook==6.4.7
notebook_shim==0.2.2
numpy==1.22.0
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.0.1
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.1
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tinycss2==1.2.1
tomli==1.2.3
toposort==1.7
tornado==6.1
tqdm==4.62.3
traitlets==5.2.0
typing_extensions==4.0.1
tzdata==2022.1
uri-template==1.2.0
urllib3==1.26.8
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
zipp==3.11.0
